Refactored custom assertions into a custom hooks setup method
